requirements.txt
torch>=2.0.0
torchvision>=0.15.0
git+https://github.com/openai/CLIP.git
streamlit>=1.28.0
numpy>=1.24.0
Pillow>=10.0.0
scipy>=1.11.0
matplotlib>=3.7.0
openpyxl>=3.0.0
tqdm>=4.65.0
